requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-redis==6.0.0
dotenv==0.9.9
google-auth==2.43.0
google-genai==1.50.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
psycopg==3.2.12
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dotenv==1.2.1
rdflib==7.4.0
redis==7.0.1
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
SPARQLWrapper==2.0.0
sqlparse==0.5.3
tenacity==9.1.2
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Wikipedia-API==0.8.1
zope.interface==8.1
